This commit implements some error handling as also it implements some logic functionality fixed, implements a searching funciton

+++ TaskManager.py
@@ -70,6 +70,7 @@ class TaskManager:
         for task in self.tasks:
             if task.taskid == id:
                 return task
+        return None
     
         
 def clearTerminal():
@@ -106,30 +107,33 @@ def ModifyTask(taskmanager):
     clearTerminal()
     print("...EDIT A TASK...")
     mytask = taskmanager.getId(id)
-    if mytask.priority == 1:
-        priority = "High"
-    elif mytask.priority == 2:
-        priority = "Medium"
+    if mytask != None:
+        if mytask.priority == 1:
+            priority = "High"
+        elif mytask.priority == 2:
+            priority = "Medium"
+        else:
+            priority = "Low"
+        print(f"Task to be edited: Name: {mytask.name}, Description: {mytask.description}, Priority: {priority}\n")
+        print("       -----------------------------------------------------------           \n")
+        option = int(input("1.Name\n2.Description\n3.Priority\n4.Cancel\nChoose an option from the menu using the numbers: ")) 
+        match option:
+            case 1:
+                newdata = input(f"Name:{mytask.name}\nWrite the new name: ")
+                mytask.name = newdata
+            case 2:
+                newdata = input(f"Description:{mytask.description}\nWrite the new description: ")
+                mytask.description = newdata
+            case 3:
+                newdata = int(input(f"Priority:{priority}\nWrite the new priority using numbers only(1.High, 2.Medium, 3.Low): "))
+                mytask.priority = newdata
+            case 4:
+                print("Canceling...")
+            case _:
+                print("This option is not available, please choose one from the menu.")
+        print("Task edited correctly.")
     else:
-        priority = "Low"
-    print(f"Task to be edited: Name: {mytask.name}, Description: {mytask.description}, Priority: {priority}\n")
-    print("       -----------------------------------------------------------           \n")
-    option = int(input("1.Name\n2.Description\n3.Priority\n4.Cancel\nChoose an option from the menu using the numbers: ")) 
-    match option:
-        case 1:
-            newdata = input(f"Name:{mytask.name}\nWrite the new name: ")
-            mytask.name = newdata
-        case 2:
-            newdata = input(f"Description:{mytask.description}\nWrite the new description: ")
-            mytask.description = newdata
-        case 3:
-            newdata = int(input(f"Priority:{priority}\nWrite the new priority using numbers only(1.High, 2.Medium, 3.Low): "))
-            mytask.priority = newdata
-        case 4:
-            print("Canceling...")
-        case _:
-            print("This option is not available, please choose one from the menu.")
-    print("Task edited correctly.")
+        print("The task was not found")
             
     
         
@@ -139,20 +143,69 @@ def DeleteTask(taskmanager):
     clearTerminal()
     print("...DELET A TASK...")
     task = taskmanager.getId(id)
-    confirm = input(f"Are you sure you want to delete ({task.name})? use y/n to yes or no: ")
-    if confirm == "y":
-        taskmanager.deleteTask(taskmanager.tasks.index(task))
-        print("Task deleted correctly.")
-    else:
-        print("Operation canceled.")
+    if task != None:
+        confirm = input(f"Are you sure you want to delete ({task.name})? use y/n to yes or no: ")
+        if confirm == "y":
+            taskmanager.deleteTask(taskmanager.tasks.index(task))
+            print("Task deleted correctly.")
+        else:
+            print("Operation canceled.")
+    else: 
+        print("Task not found.")
     
 
 def MarkCompleted(taskmanager):
     print("...MARK A TASK COMPLETED...")
     id = int(input("Write the id of the task to be marked as completed: "))
     task = taskmanager.getId(id)
-    taskmanager.taskCompleted(taskmanager.tasks.index(task))
-    print("Task marked correctly correctly.")
+    if task != None:
+        taskmanager.taskCompleted(taskmanager.tasks.index(task))
+        print("Task marked correctly correctly.")
+    else:
+        print("The task was not found.")
+
+def SearchTask(taskmanager):
+    print("...SEARCH FOR A TASK...")
+    option = int(input("Optoins:\n1.Search for ID\n2.Search for name or letter\n Select an option using numbers: "))
+    clearTerminal()
+    print("...SEARCH FOR A TASK...\n\n")
+    print("   ------------------------    \n")
+    
+    match option:
+        case 1:
+            id = int(input("Write the ID for the task: "))
+            task = taskmanager.getId(id)
+            if task != None:
+                match task.priority:
+                    case 1:
+                        priority = "High"
+                    case 2:
+                        priority = "Medium"
+                    case 3: priority = "Low"
+                if task.state == 1:
+                    state = "Completed"
+                else:
+                    state = "Pending"
+                print(f"{task.taskid}. {task.name}\n Description:\n{task.description}\n\nPriority: {priority} State: {state}")
+            else:
+                print("No task was found.")
+        case 2:
+            search = input("Write the name or wordsd that match with the name of the task: ")
+            mylist = []
+            print(f"Matches for {search}.\n")
+            print("  ID  |    Name    ")
+            for task in taskmanager.tasks:
+                if search.lower() in task.name.lower():
+                    print(f" {task.taskid}          {task.name}")
+                    mylist.append(task)
+            if len(mylist) == 0:
+                print(f"No tasks where found using {search}")
+        case _:
+            print("The option selected is not in the menu, try again.")
+            
+    input("Press enter to continue...")
+        
+
 
 def menu():
     print("TASK MANAGER")
@@ -161,7 +214,8 @@ def menu():
     print("3.Delete a Task")        
     print("4.Mark as completed")
     print("5.Show all Tasks")
-    print("6.Exit")
+    print("6.Search a task")
+    print("7.Exit")
     option = int(input("Choose an option from the menu using the numbers: "))
     return option
 
@@ -215,13 +269,18 @@ def main(mytaskmanager: TaskManager):
                 mytaskmanager.showTasks()
                 input("\n\nPress any key to go to menu... ")
             case 6:
+                clearTerminal()
+                SearchTask(mytaskmanager)
+                time.sleep(2)
+            case 7:
                 print("Exiting....")
                 key = False
                 time.sleep(2)
                 clearTerminal()
             case _:
                 print("Sorry this option is not available.")
-    writeJson(mytaskmanager)
+        writeJson(mytaskmanager)
+    
         
         
 clearTerminal()
